refactor: move classify_event and classify_error into ExecutionEvent
These functions operate solely on ExecutionEvent data — moving them there
eliminates Feature Envy and keeps domain semantics co-located with the type.

Callers (TaskChannel, SwarmDispatcher) now call ExecutionEvent.classify/1
and ExecutionEvent.classify_error/1 directly.
